selgo-backend/square-service/requirements.txt
alembic==1.10.3
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.4.26
click==8.2.0
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.95.1
GeoAlchemy2==0.13.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.6.1
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.13.2
uvicorn==0.21.1
